Update standard_transaction.md
The use of the word "original" in the current version of our first proposed change can be confused with the previous use earlier in the text, where it referred to the original public key, i.e. the public key belonging to the person who can do a delegated spend.

Added another sentence to help provide more clarity about the secret component of the synthetic private key, also.